refactor: remove unused Alignment Guides component
