+++ frontend/src/components/SnapGuides.vue
